requirements.txt
aiohttp
asyncpg
hikari
hikari-lightbulb
psutil
python-dotenv
uvloop